Started random forest/decision tree
I think my basic implementation is correct.  May want to make a more
complicated decision tree to use, as it just chooses a random attribute and
splits at the moment.  Does really badly on dna (~53%) and ocr49 (~63%), but OK
on census (80%) and ocr17 (90%).  All of these are with 1000 trees.
DecisionForest takes the size of the forest as the second argument.