More work on the decision tree.  I honestly do not know why it is doing so badly.
I suspect that my gain/entropy calculations are not correct (see around line 100).
Both files have a basic cross validator that take the last 1/4 of the dataset and
use it to check training on the first 3/4.  I will probably change it to randomize
the set later, but for now I am just trying to get it to work.