...Fixed it.  I am a dumbass. See below.
So all of my splitting and attribute selection was fine and dandy.  The problem
was that when I split off examples to put in the 0 or 1 branch, I was using the
index i in the examples vector rather than the value at index i in the example
vector (line 153 in DecisionTree).  Performance is now really good for the
decision tree.  I'm still going to convert to discrete values, though.