Discrete values in decision tree / testing harness
I changed the dataset for decision tree from binary to discrete.  It gives
about the same results, but it seems to be significantly faster (I think this
makes sense - we split the examples more often).  Also created a testing
harness that takes the filename, algorithm, and number of trials.  It trains on
part of the dataset and tests on the other part.